refactor: Row as own component
